Create UrduWordList
This word list consists of unique words taken from Urdu hun spell list created on Urduweb.org/mehfil and Jameel Noori Nastaleeq 3 ligatures created by the same community. Additionally, Urdu special characters including diacritic marks and superscript characters are also included at the and of this list. The purpose of this file is to provide a base to create concordance lines for training text and train Tesseract 4 OCR for Urdu.